+++ frontend/wailsjs/go/app/App.d.ts
@@ -20,6 +20,8 @@ export function DropDatabase(arg1:string,arg2:string):Promise<boolean>;
 
 export function DropIndex(arg1:string,arg2:string,arg3:string,arg4:string):Promise<boolean>;
 
+export function DuplicateCollection(arg1:string,arg2:string,arg3:string,arg4:string,arg5:string,arg6:string):Promise<boolean>;
+
 export function Environment():Promise<app.EnvironmentInfo>;
 
 export function ExecuteShellScript(arg1:string,arg2:string,arg3:string,arg4:string):Promise<app.ExecuteShellScriptResult>;