+++ frontend/wailsjs/go/app/App.js
@@ -30,6 +30,10 @@ export function DropIndex(arg1, arg2, arg3, arg4) {
   return window['go']['app']['App']['DropIndex'](arg1, arg2, arg3, arg4);
 }
 
+export function DuplicateCollection(arg1, arg2, arg3, arg4, arg5, arg6) {
+  return window['go']['app']['App']['DuplicateCollection'](arg1, arg2, arg3, arg4, arg5, arg6);
+}
+
 export function Environment() {
   return window['go']['app']['App']['Environment']();
 }